Add database implementation
So, this isn't going to work how we though. The database code either needs to incorporate callbacks or promises. What has been implemented here is horrid and needs to be completely redone. The code Kevin wrote could not be implemented in any functional way, but designs have been copied over where possible. To be discussed Sunday meeting.